requirements.txt
loguru >= 0.7.3
python-dotenv >= 1.2.1
aiogram >= 3.24.0